refactor(dashboard): hoist formatRelativeTime into the shared activity format utils
