+++ apps/sdp-web/src/app/dashboard/activity-format-utils.ts
@@ -7,6 +7,29 @@ export function toTitleCase(value: string): string {
     .join(" ");
 }
 
+export function formatRelativeTime(value: string): string {
+  const date = new Date(value);
+  if (Number.isNaN(date.getTime())) {
+    return "—";
+  }
+
+  const diffMs = date.getTime() - Date.now();
+  const diffMinutes = Math.round(diffMs / 60_000);
+  const formatter = new Intl.RelativeTimeFormat("en-US", { numeric: "auto" });
+
+  if (Math.abs(diffMinutes) < 60) {
+    return formatter.format(diffMinutes, "minute");
+  }
+
+  const diffHours = Math.round(diffMinutes / 60);
+  if (Math.abs(diffHours) < 24) {
+    return formatter.format(diffHours, "hour");
+  }
+
+  const diffDays = Math.round(diffHours / 24);
+  return formatter.format(diffDays, "day");
+}
+
 export function readTransactionParam(
   params: Record<string, unknown>,
   key: string
